Изменил название пакета с entities на model
