+++ src/main/java/io/ylab/ticTacToeGameRestApi/model/GameStatus.java
@@ -1,4 +1,4 @@
-package io.ylab.ticTacToeGameRestApi.entities;
+package io.ylab.ticTacToeGameRestApi.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,35 +8,27 @@ import lombok.Setter;
 import javax.persistence.*;
 
 @Entity
-@Table(name = "steps")
+@Table(name = "game_status")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Step {
+public class GameStatus {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
 
-    @Column(name = "num")
-    private Integer num;
-
-    @Column(name = "symbol")
-    private String symbol;
-
-    @Column(name = "column")
-    private Integer column;
-
-    @Column(name = "line")
-    private Integer row;
-
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "player_id")
-    private Player player;
-
     @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "game_id")
     private Game game;
+
+    @Column(name = "status")
+    private String status;
+
+    public GameStatus(Game game, String status) {
+        this.game = game;
+        this.status = status;
+    }
 }
