+++ src/main/java/io/ylab/ticTacToeGameRestApi/repositories/GameRepository.java
@@ -1,6 +1,6 @@
 package io.ylab.ticTacToeGameRestApi.repositories;
 
-import io.ylab.ticTacToeGameRestApi.entities.Game;
+import io.ylab.ticTacToeGameRestApi.model.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
