+++ src/main/java/io/ylab/ticTacToeGameRestApi/repositories/GameResultRepository.java
@@ -1,6 +1,6 @@
 package io.ylab.ticTacToeGameRestApi.repositories;
 
-import io.ylab.ticTacToeGameRestApi.entities.GameResult;
+import io.ylab.ticTacToeGameRestApi.model.GameResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
