+++ src/main/java/io/ylab/ticTacToeGameRestApi/repositories/GameplayRepository.java
@@ -1,6 +1,6 @@
 package io.ylab.ticTacToeGameRestApi.repositories;
 
-import io.ylab.ticTacToeGameRestApi.entities.Gameplay;
+import io.ylab.ticTacToeGameRestApi.model.Gameplay;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
