+++ src/main/java/io/ylab/ticTacToeGameRestApi/services/GameResultService.java
@@ -1,7 +1,7 @@
 package io.ylab.ticTacToeGameRestApi.services;
 
-import io.ylab.ticTacToeGameRestApi.entities.GameResult;
-import io.ylab.ticTacToeGameRestApi.entities.Player;
+import io.ylab.ticTacToeGameRestApi.model.GameResult;
+import io.ylab.ticTacToeGameRestApi.model.Player;
 
 public interface GameResultService {
