+++ src/main/java/io/ylab/ticTacToeGameRestApi/services/serviceImp/GameResultServiceImp.java
@@ -1,7 +1,7 @@
 package io.ylab.ticTacToeGameRestApi.services.serviceImp;
 
-import io.ylab.ticTacToeGameRestApi.entities.GameResult;
-import io.ylab.ticTacToeGameRestApi.entities.Player;
+import io.ylab.ticTacToeGameRestApi.model.GameResult;
+import io.ylab.ticTacToeGameRestApi.model.Player;
 import io.ylab.ticTacToeGameRestApi.repositories.GameResultRepository;
 import io.ylab.ticTacToeGameRestApi.services.GameResultService;
 import lombok.AllArgsConstructor;
